+++ packages/cli/CHANGELOG.md
@@ -9,9 +9,9 @@
 ### Patch Changes
 
 - Updated dependencies
-  - @kenkaiiii/gg-ai@4.12.0
-  - @kenkaiiii/gg-agent@4.12.0
-  - @kenkaiiii/gg-core@4.12.0
+  - @prestyj/ai@4.12.0
+  - @prestyj/agent@4.12.0
+  - @prestyj/core@4.12.0
 
 ## 4.11.3